Move PayPalApprovalURLParser into own file
