+++ Sources/BraintreePayPal/BTPayPalClient.swift
@@ -297,35 +297,6 @@ import BraintreeDataCollector
         }
     }
     
-    enum PayPalRedirectType {
-        case webBrowser(url: URL)
-        case payPalApp(url: URL)
-    }
-    
-    struct ApprovalURLParser {
-        
-        var redirectType: PayPalRedirectType
-        
-        var pairingID: String? {
-            switch redirectType {
-            case .webBrowser(let url), .payPalApp(let url):
-                let url = URLComponents(url: url, resolvingAgainstBaseURL: true)
-                return url?.queryItems?.first(where: { $0.name == "token" || $0.name == "ba_token" })?.value
-            }
-        }
-        
-        init?(body: BTJSON) {
-            if let paypalAppRedirectUrl = body["paymentResource"]["paypalAppApprovalUrl"].asURL() {
-                redirectType = .payPalApp(url: paypalAppRedirectUrl)
-            } else if let approvalURL = body["paymentResource"]["redirectUrl"].asURL() ??
-                body["agreementSetup"]["approvalUrl"].asURL() {
-                redirectType = .webBrowser(url: approvalURL)
-            } else {
-                return nil
-            }
-        }
-    }
-    
     private func launchPayPalApp(with paypalAppRedirectUrl: URL, completion: @escaping (BTPayPalAccountNonce?, Error?) -> Void) {
         var urlComponents = URLComponents(url: paypalAppRedirectUrl, resolvingAgainstBaseURL: true)
         urlComponents?.queryItems = [
@@ -340,7 +311,6 @@ import BraintreeDataCollector
 
         application.open(redirectURL, options: [:]) { success in
             // TODO: - Handle success or fail of opening app
-            completion(BTPayPalAccountNonce(json: BTJSON()), nil)
         }
     }
